Move user reducer to a new file
